# Performance improvement for QR rendering