--- a/src/pdf/pdf.ts
+++ b/src/pdf/pdf.ts
@@ -549,16 +549,11 @@ export class PDF_ extends ExtendedPDF {
 
   private _renderQRCode(): void {
 
-    const qrcode = generateQRCode(this._data, utils.mm2pt(46));
-
+    const qrcode = generateQRCode(this._data, "pdf", utils.mm2pt(67), this._marginTop + utils.mm2pt(17), utils.mm2pt(46));
 
     //-- Add QR Code
 
-    this.addPath(qrcode, utils.mm2pt(67), this._marginTop + utils.mm2pt(17))
-      .undash()
-      .fillColor("black")
-      .fill();
-
+    this.fillColor("black").addContent(qrcode).fill();
 
     //-- Add Swiss Cross