--- a/src/shared/qr-code.ts
+++ b/src/shared/qr-code.ts
@@ -2,7 +2,15 @@ import { getReferenceType } from "../shared/utils.js";
 import { Data } from "./types";
 import { qrcodegen } from "./qr-code-generator.js";
 
-export default function generateQRCode(data: Data, size: number): string {
+function number(n: number) {
+  if(n > -1e21 && n < 1e21){
+    return Math.round(n * 1e6) / 1e6;
+  }
+
+  throw new Error(`unsupported number: ${n}`);
+}
+
+export default function generateQRCode(data: Data, type: "pdf" | "svg", xOrigin: number, yOrigin: number, size: number): string {
 
   let qrString = "";
 
@@ -238,7 +246,17 @@ export default function generateQRCode(data: Data, size: number): string {
     for(let y = 0; y < qrCode.size; y++){
       const yPos = y * blockSize;
       if(qrCode.getModule(x, y)){
-        parts.push(`M ${xPos}, ${yPos} V ${yPos + blockSize} H ${xPos + blockSize} V ${yPos} H ${xPos} Z `);
+
+        switch (type){
+          case "pdf":
+            parts.push(`${number(xOrigin + xPos)} ${number(yOrigin + yPos)} ${number(blockSize)} ${number(blockSize)} re`);
+            break;
+
+          case "svg":
+            parts.push(`M ${xPos}, ${yPos} V ${yPos + blockSize} H ${xPos + blockSize} V ${yPos} H ${xPos} Z `);
+            break;
+        }
+
       }
     }