It takes a lot of time to render a lot of QR Bills since addPath is an expensive function as we need to convert from QR segments to SVG and then to PDF.

We can bypass the conversion from QR to SVG, and directly convert QR segments to PDF.

This leads to huge performance benefits (a4.js):
| Number of bills | 10        | 100       | 1000    | 10000           |
|-----------------|-----------|-----------|---------|-----------------|
| Normal          | 754.651ms | 5.426s    | 50.278s | Node crashed :( |
| Modified        | 278.702ms | 969.642ms | 6.985s  | 1 min 7.421s    |