@danielpanero 2022-08-28
Furthermore, the second most expensive function is the `getPenaltyScore()` found in qr-code-generator.ts, that checks which of the 8 mask patterns is the most effective one...

`const qrCode = qrcodegen.QrCode.encodeSegments([eci, ...segments], qrcodegen.QrCode.Ecc.MEDIUM, 1, 40, 0);`

If we impose mask 0, this leads to again to huge performance gains:
| Number of bills | 10        | 100       | 1000    | 10000           |
|-----------------|-----------|-----------|---------|-----------------|
| Normal          | 754.651ms | 5.426s    | 50.278s | Node crashed :( |
| Modified        | 278.702ms | 969.642ms | 6.985s  | 1 min 7.421s    |
| Modified V2     | 157.408ms | 586.346ms | 3.679s  | 34.487s         |