@schoero 2022-08-29
Thank you for further researching and testing performance bottlenecks. It sounds good, but I'm not sure if it's the best idea to hardcode the masking pattern. After all, the readability of the QR code will suffer in most cases. Do you have a more in-depth knowledge about QR codes and do you know if this is safe to do?

I don't really want to trade readability for performance.