@danielpanero 2022-08-31
Yes, that's the problem... for 10/100 bills the performance gains are not justifiable as I think it's better to lose some time improving the readability of the code. After 1000 I don't know...