@schoero 2022-08-31
I checked the [specifications](https://www.paymentstandards.ch/dam/downloads/ig-qr-bill-en.pdf) again. There is no mention about the masking patterns. But there are rules to minimum module size and error correction etc, so I would like to play it save and stick to the expensive calculation for now. 

Im going to merge, feel free to open an issue if you want to further discuss this topic.